+++ gpu-video/src/global_registry.rs
@@ -4,7 +4,7 @@ use ash::vk;
 use rustc_hash::FxHashMap;
 use wgpu::hal::vulkan::Api as VkApi;
 
-use crate::VideoDevice;
+use crate::device::VideoDevice;
 
 #[derive(Default)]
 pub(crate) struct GlobalRegistry {
@@ -15,11 +15,11 @@ static REGISTRY: LazyLock<RwLock<GlobalRegistry>> =
     LazyLock::new(|| RwLock::new(GlobalRegistry::default()));
 
 impl GlobalRegistry {
-    pub(crate) fn register_device(handle: VideoDeviceKey, device: Arc<VideoDevice>) {
+    pub(crate) fn register_device(key: VideoDeviceKey, device: Arc<VideoDevice>) {
         let mut registry = REGISTRY.write().unwrap();
 
         use std::collections::hash_map::Entry;
-        match registry.devices.entry(handle) {
+        match registry.devices.entry(key) {
             Entry::Occupied(_) => {
                 tracing::debug!("Tried to register device that already exists in the registry");
             }
@@ -29,18 +29,18 @@ impl GlobalRegistry {
         }
     }
 
-    pub(crate) fn unregister_device(handle: &VideoDeviceKey) {
+    pub(crate) fn unregister_device(key: &VideoDeviceKey) {
         let mut registry = REGISTRY.write().unwrap();
-        if registry.devices.remove(handle).is_none() {
+        if registry.devices.remove(key).is_none() {
             tracing::debug!("Tried to unregister device that does not exist in the registry");
         }
     }
 
-    pub(crate) fn get_device(handle: &VideoDeviceKey) -> Result<Arc<VideoDevice>, RegistryError> {
+    pub(crate) fn get_device(key: &VideoDeviceKey) -> Result<Arc<VideoDevice>, RegistryError> {
         let registry = REGISTRY.read().unwrap();
         registry
             .devices
-            .get(handle)
+            .get(key)
             .cloned()
             .ok_or(RegistryError::DeviceNotFound)
     }
@@ -49,13 +49,19 @@ impl GlobalRegistry {
 // TODO: metal key
 #[cfg(vulkan)]
 #[derive(Debug, Copy, Clone, Hash, PartialEq, Eq)]
-pub(crate) struct VideoDeviceKey(vk::Device, vk::Queue);
+pub(crate) struct VideoDeviceKey(pub(crate) vk::Device, pub(crate) vk::Queue);
 
 #[cfg(vulkan)]
 impl From<&wgpu::Device> for VideoDeviceKey {
     fn from(device: &wgpu::Device) -> Self {
         let hal_device = unsafe { device.as_hal::<VkApi>().unwrap() };
         Self(hal_device.raw_device().handle(), hal_device.raw_queue())
+    }
+}
+
+impl From<&crate::VideoDevice> for VideoDeviceKey {
+    fn from(device: &crate::VideoDevice) -> Self {
+        device.handle.0
     }
 }
